only use CMPFN in check-valid-sort
`check-valid-sort` allows items that are equal in a series, but CMPFN must be
a strict ordering predicate (e.g. less-than instead of less-than-or-equal).
Using `equal?` does not work for all data types, in particular `(equal? 0
0.0)` is false.

Changed the criteria to only use CMPFN to check that a series is sorted.  Two
items are equal if both (CMPFN a b) and (CMPFN b a) are false.